+++ playlist_bridge/__init__.py
@@ -1,5 +1,5 @@
 """Playlist Bridge package."""
 
-__version__ = "2.1.0-beta.2"
+__version__ = "2.1.0-beta.3"
 
-__build__ = "20260916.3"
+__build__ = "20260916.4"